requirements.txt
# AI Trading Dashboard Dependencies
# Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data & ML
yfinance==0.2.28
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2

# Database
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Task Queue
celery==5.3.4

# API & WebSocket
requests==2.31.0
websockets==12.0
python-socketio==5.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
